REFACTOR: 패키지 변경
- 도메인 레이어가 서비스 레이어를 의존함.
