+++ src/main/java/com/athenhub/productservice/product/domain/repository/ProductDetailRepository.java
@@ -1,8 +1,8 @@
 package com.athenhub.productservice.product.domain.repository;
 
-import com.athenhub.productservice.product.application.dto.SearchProductResponse;
 import com.athenhub.productservice.product.domain.Product;
 import com.athenhub.productservice.product.domain.dto.SearchDaoRequest;
+import com.athenhub.productservice.product.domain.dto.SearchProductResponse;
 import com.athenhub.productservice.product.domain.vo.HubId;
 import com.athenhub.productservice.product.domain.vo.ProductVariantId;
 import com.athenhub.productservice.product.domain.vo.VendorId;
