+++ src/main/java/com/athenhub/productservice/product/presentation/controller/ProductQueryController.java
@@ -5,11 +5,11 @@ import com.athenhub.productservice.membership.domain.MemberRole;
 import com.athenhub.productservice.membership.domain.MemberRoles;
 import com.athenhub.productservice.product.application.dto.ProductDetail;
 import com.athenhub.productservice.product.application.dto.ProductSummary;
-import com.athenhub.productservice.product.application.dto.SearchProductResponse;
 import com.athenhub.productservice.product.application.service.ProductQueryApplicationService;
 import com.athenhub.productservice.product.application.service.ProductQueryService;
 import com.athenhub.productservice.product.domain.Product;
 import com.athenhub.productservice.product.domain.dto.SearchDaoRequest;
+import com.athenhub.productservice.product.domain.dto.SearchProductResponse;
 import java.util.List;
 import java.util.UUID;
 import java.util.stream.Stream;
